Loadcache

DataToolkitCommon.LOADCACHE_PLUGIN — Constant

Cache the results of data loaders using JLD2. Cache file paths are determined by the dataset UUID and a hash of the loader and storage backends.

It is important to note that this comes with the limitations of JLD2, i.e. some data types will not be able to be cachced effectively, such as IOStream.

Enabling caching of a loader

To cache the result of an individual loader, set the cache parameter.

[[mydata.loader]]
driver = "backend"
cache = true

To apply this setting in bulk, the defaults plugin may be of interest. With it, all loaders can be cached with the following:

[config.defaults.loader._]
cache = true

Configuring the cache directory

The cache directory can be set globally, or per-loader.

[config.loadcache]
folder = "loadcache" # the default

[somedata.loader]
loadcache = "some_folder"