Store

DataToolkitCommon.Store.STORE_PLUGIN — Constant

Cache IO from data storage backends, by saving the contents to the disk.

Configuration

Saving of individual storage sources can be disabled by setting the "save" parameter to false, i.e.

[[somedata.storage]]
save = false
...

System-wide configuration can be set via the store config set REPL command, or directly modifying the DataToolkitCommon.Store.INVENTORY.config struct.

Four (system-wide) settings determine garbage collection behaviour:

auto_gc (default 2): How often to automatically run garbage collection (in hours). Set to a non-positive value to disable.

max_age (default 30): The maximum number of days since a collection was last seen before it is removed from consideration.

max_size (default 53687091200): The maximum (total) size of the store.

recency_beta (default 1): When removing items to avoid going over max_size, how much recency should be valued. Can be set to any value in (-∞, ∞). Larger (positive) values weight recency more, and negative values weight size more. -1 and 1 are equivalent.